Update : 인터페이스 분리 수정
