+++ src/main/java/com/project/devidea/modules/account/AccountController.java
@@ -3,7 +3,7 @@ package com.project.devidea.modules.account;
 import com.project.devidea.infra.config.security.LoginUser;
 import com.project.devidea.infra.error.GlobalResponse;
 import com.project.devidea.modules.account.dto.*;
-import com.project.devidea.modules.account.service.AccountServiceImpl;
+import com.project.devidea.modules.account.service.AccountService;
 import com.project.devidea.modules.account.validator.SignUpOAuthRequestValidator;
 import com.project.devidea.modules.account.validator.SignUpRequestValidator;
 import io.swagger.annotations.ApiOperation;
@@ -25,7 +25,7 @@ import java.util.Map;
 @RequiredArgsConstructor
 public class AccountController {
 
-    private final AccountServiceImpl accountServiceImpl;
+    private final AccountService accountService;
     private final SignUpRequestValidator signUpRequestValidator;
     private final SignUpOAuthRequestValidator signUpOAuthRequestValidator;
 
@@ -43,14 +43,14 @@ public class AccountController {
     @ApiOperation("회원가입")
     public ResponseEntity<?> signUp(@Valid @RequestBody SignUp.CommonRequest commonRequest) {
 
-        return new ResponseEntity<>(GlobalResponse.of(accountServiceImpl.signUp(commonRequest)), HttpStatus.OK);
+        return new ResponseEntity<>(GlobalResponse.of(accountService.signUp(commonRequest)), HttpStatus.OK);
     }
 
     @GetMapping("/authenticate-email-token")
     @ApiOperation("일반 회원가입 메일 인증")
     public void authenticateEmailToken(String email, String token, HttpServletResponse response) throws IOException {
 
-        String url = accountServiceImpl.authenticateEmailToken(email, token);
+        String url = accountService.authenticateEmailToken(email, token);
         response.sendRedirect(url + "/sign-up/detail?token=" + token);
     }
 
@@ -59,14 +59,14 @@ public class AccountController {
     public ResponseEntity<?> signUpOAuth(@Valid @RequestBody SignUp.OAuthRequest oAuthRequest)
             throws NoSuchAlgorithmException {
 
-        return new ResponseEntity<>(GlobalResponse.of(accountServiceImpl.signUpOAuth(oAuthRequest)), HttpStatus.OK);
+        return new ResponseEntity<>(GlobalResponse.of(accountService.signUpOAuth(oAuthRequest)), HttpStatus.OK);
     }
 
     @PostMapping("/login")
     @ApiOperation("로그인")
     public ResponseEntity<?> login(@Valid @RequestBody Login.Common login) throws Exception {
 
-        Map<String, String> result = accountServiceImpl.login(login);
+        Map<String, String> result = accountService.login(login);
         return new ResponseEntity<>(GlobalResponse.of(getIsSavedDetails(result)),
                 getHttpHeaders(result), HttpStatus.OK);
     }
@@ -75,7 +75,7 @@ public class AccountController {
     @ApiOperation("로그인 - OAuth")
     public ResponseEntity<?> loginOAuth(@Valid @RequestBody Login.OAuth login) throws Exception {
         // 디테일 입력했는지 여부 확인하기
-        Map<String, String> result = accountServiceImpl.loginOAuth(login);
+        Map<String, String> result = accountService.loginOAuth(login);
         return new ResponseEntity<>(GlobalResponse.of(getIsSavedDetails(result)),
                 getHttpHeaders(result), HttpStatus.OK);
     }
@@ -96,7 +96,7 @@ public class AccountController {
     @ApiOperation("회원가입 디테일")
     public ResponseEntity<?> signUpDetail(@Valid @RequestBody SignUp.DetailRequest detailRequest) {
 
-        accountServiceImpl.saveSignUpDetail(detailRequest);
+        accountService.saveSignUpDetail(detailRequest);
         return new ResponseEntity<>(GlobalResponse.of(), HttpStatus.OK);
     }
 
@@ -104,7 +104,7 @@ public class AccountController {
     @ApiOperation("회원탈퇴")
     public ResponseEntity<?> quit(@AuthenticationPrincipal LoginUser loginUser) {
 
-        accountServiceImpl.quit(loginUser);
+        accountService.quit(loginUser);
         return new ResponseEntity<>(GlobalResponse.of(), HttpStatus.OK);
     }
 }
