+++ src/main/java/com/project/devidea/modules/account/service/AccountServiceImpl.java
@@ -30,14 +30,13 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.security.NoSuchAlgorithmException;
-import java.time.LocalDateTime;
 import java.util.*;
 import java.util.stream.Collectors;
 
 @Service
 @Transactional
 @RequiredArgsConstructor
-public class AccountServiceImpl implements OAuthService, AccountService, AccountInfoService {
+public class AccountServiceImpl implements AccountService {
 
     private final BCryptPasswordEncoder passwordEncoder;
     private final AccountRepository accountRepository;
